Add stack side interface for Bluetooth audio HAL V2
This bases on the new Bluetooth Audio HAL V2 to provide a HIDL based
interface for the stack. There will be a common interface named
BluetoothAudioClientInterface that the stack can register its session for
stream of A2DP or Hearing Aid. When the stack registers to this
pre-implemented BluetoothAudioClientInterface, the audio HAL can control
the stream state and the stack can report results to the audio HAL.
When running for software encoding, there will also data path via FMQ to
provide the bridge between audio HAL and the stack. This change contains
A2DP software encoding (legacy) only.

Bug: 111519504
Test: manual with A2DP software encoding (legacy)

Change-Id: Iac5a43c929d4036fa86e2b0c2c2920ca2b9dfa50