Enable shadow call stack in system/bt.
Shadow call stack (SCS) is a security mitigation that uses a separate
stack (the SCS) for return addresses.

This change causes libbluetooth-binder.so and libbluetooth.so to be
built with SCS.

For the moment this is an internal-only change until the AOSP kernels
can be fixed to not clobber x18.

Bug: 112907825
Bug: 120630739
Change-Id: I0b9e948f26b4430dd04f44ffc94abf96d2a7ee8f