hf_client: Pass an empty string when number_present is false
When the AG dials a number that is more than 32 digits long, the CLCC
response to HF contains an invalid number.  This will be detected and
eventually leads to the number string being null when calling isValidUtf
in current_calls_cb.  This causes isValidUtf() to crash with a null pointer
dereference.

Bug: 128910739
Bug: 109838537

Test: local test to observe no crash at HF side when dialing a long
invalid number in AG (paired phone).

Change-Id: Ie1b9f62007d523e644b54bda50f11975f2b3d034