Call gatt_disconnect if the LE connection is not established and no APP use it.
When the LE connection hold APP is removed, if the LE connection is not
established and no APP use it, we never free transport control block
(p_tcb) or close it. For BLE device that have not established and no APP
use it, will call gatt_disconnect to free p_tcb and go to close state.

Bug: 127908699
Test: GattConnectTest
Change-Id: Ifed10c67141cb4c9221478db759c5614714fff30