Rootcanal HCI HAL: Set read from socket blocking
In HciSocketDevice::SendHci, rootcanal first writes to HCI socket one
byte HCI packet type first, then HCI packet body. When the stack
receives the first byte, it tries to read the entire packet, but
sometimes the rest of packet was not delivered yet, so the stack will
crash. Instead, the stack should be blocked until all bytes are read,
with a 3 second timeout when rootcanal can't deliver all bytes to stack.

Test: try cert/run_cert.sh multiple times
Bug: 131625008
Change-Id: I99df53c658717a5c931ebc075eb75df447e10332