Hearing Aid: properly track audio Suspend/Resume
When audio is suspended or resumed, we are notified by OnAudioSuspend or
OnAudioResume callback. Unfortunately, when we are the one triggering
the HearingAidAudioSource::Stop(), we would not receive the suspend
callback. This is causing HA to believe audio is being streamed when
it's not.

This patch fixes that by setting audio_running to false when we make
call to Stop()

Bug: 132793574
Test: Play audio, restart Hearing Aid, verify START command is not being
sent

Change-Id: I3c4551114ac275ea8c72f64a8d036373429acd6d