Refresh the encoder configuration of software while updating the MTU
In order to have better compatibility, peer parameters like the MTU are
necessary for packing AVDTP media frames. When reconfiguring the codec,
the encoder was updated before negotiated with remote, and needed to
update again after data channel opened. There was set audio config from
UIPC to help, and now we add the request after reconfigured.

Bug: 134125165
Test: manual
Change-Id: Ia823057b13cebde63d72f2a41aa744f1966eb20d