A2DP: Clear the collision mask while disconnecting
While remote is opening AVDTP channel, the stack setup a coll_mask
flags with a signalling timer, and takes over the control after timer
expired. This prevents the connection collision. However, if there were
two incoming connections at the same time, the timer was overridden by
the second connection, and this flags of the first connection kept
activated, but the following outgoing connection was blocked. In order
to prevent the worng flags, this change reset flags while disconnecting.

Bug: 146849922
Test: Connect with multiple A2DP sinks
Change-Id: I4baff647d22f1148bf6fa2b61e778652c9c90091