L2cap cert tests: Set initator address from the test
Add an API to AclManager to set the initiator address in tests.

This CL fixes issue only in tests that use _setup_link_from_cert, and
not use _set_link_from_dut_and_open_channel. This is because both dut
and cert used the same hardcoded address for initiation.

This will be extended once we have proper RPA.

Test: gd/cert/run --host --test_filter=LeL2capTest
Bug: 152348535
Change-Id: I96e6d23de20bbc1d5afd699908522ca672258537