L2cap Performance Test initial commit
Add the PerformanceTestLogger as a common helper.

Add the first TX/RX test for L2cap. So far the tests are based on
sending x packets (a small number) of size y each to get the rough idea
about performance and use it as a benchmark. In further iteration, we
can add tests to send packets of size y each over duration z.

Currently we log the interval through PerformanceTestLogger. In further
iteration, we need to report this to test result.

Increased the threshold for ERTM local busy detection to avoid rejecting
remote packets too frequently. So far when the performance test sends
multiple packets without receiving ack, it won't stop when it receives
RNR (DUT busy).

Tag: #gd-refactor
Bug: 157613312
Test: cert/run --host
Change-Id: If8960001473590ca5c9e2bab3bb63271cb86ce4a