Cert: Add new functionality to cert/run
Place the venv in ashmem by default.
This decreases the --clean build runtime
by ~40%.  This also decreases the regular build runtime
by ~10%.

The build time that is measured is the time for the
python/cert changes to happen.  This isn't accounting
for building C++ files and objects into intermediates
and target.

Please note, if you lose power or reboot you will lose
data in ashmem.

Added a flag '--gotta-go-slow' for those that want the venv
to remain on disk.

Added a flag '--speed-hax' as a bucket to add non default
speed improvements. Currently, this will move HostOnlyCert
to ashmem which decreases clean builds by 54% and
non clean builds by 16% when coupled
with default gotta go fast feature.

Bug: 159358844
Test: cert/run --host --verbose --test_filter=SecurityTest:test_dut_initiated_no_input_no_output_no_input_no_output
Test: cert/run --clean --host --verbose --test_filter=SecurityTest:test_dut_initiated_no_input_no_output_no_input_no_output
Test: cert/run --gotta-go-slow --host --verbose --test_filter=SecurityTest:test_dut_initiated_no_input_no_output_no_input_no_output
Test: cert/run --clean --gotta-go-slow --host --verbose --test_filter=SecurityTest:test_dut_initiated_no_input_no_output_no_input_no_output
Test: cert/run --speed-hax --host --verbose --test_filter=SecurityTest:test_dut_initiated_no_input_no_output_no_input_no_output
Test: cert/run --gotta-go-slow --speed-hax --host --verbose --test_filter=SecurityTest:test_dut_initiated_no_input_no_output_no_input_no_output
Tag: #gd-refactor
Change-Id: Ibdf42f0cf7c2e9a32e3360f2f071443722796a4f