Cert: run script updates.
This script takes a looong time to run.  This slows down
development a lot.

I had added offloading pieces to ashmem to help with speed and
this does help.  However, speedhax had a bug, and the flag seemed
confusing.  I took a step back and looked at the script as a whole.

There were 3 ways to set the test filter, one which
didn't make much sense.  We should deduplicate options
and complexity.
Option 1) --test_filter=*
Option 2) --test_file=*
Option 3) $1 (Bash argument that isn't consumed is the group/test target [e.g. Securitytest]))

I chose to use Option 3, and remove 1 & 2 since you don't need to type '--flag_name='

speedhax is non descriptive, gotta go fast is useless as a default.
so I renamed the flags and reworked the flow a bit.  This should
also fix bugs that I was seeing occasionally.

The bug was cross device linking failed.  This seemed to happen when
moving the data from disk to ashmem.  So for now with this CL
it will just remove everything and remake with the flag change.
I will investigate a move option at a later date.

Introduce '--skip-soong-build'
Soong is only needed when you need to build C/C++ targets.  A lot
of the time when I am writing cert tests I make minimal changes to
the C/C++ files.  Once I have them in place I write the test with python.
I've been using this flag to speed up my process

Introduce '--help'
this is to be updated when you add a new argument!

Bug: 159358844
Test: cert/run SecurityTest
Test: cert/run --host SecurityTest
Test: cert/run --clean --host SecurityTest
Test: cert/run --disable-ashmem-venv --host SecurityTest
Test: cert/run --disable-ashmem-venv --clean --host SecurityTest
Test: cert/run --host --verbose SecurityTest:test_dut_initiated_display_only_display_only --skip-soong-build
Test: cert/run --gotta-go-fast --host --verbose SecurityTest:test_dut_initiated_display_only_display_only --skip-soong-build
Test: cert/run --clean --gotta-go-fast --host --verbose SecurityTest:test_dut_initiated_display_only_display_only --skip-soong-build --repeat=2

Tag: #gd-refactor
Change-Id: Ib2a978900988bec01654e853d515da1175720d32