Shim: Minor fix to bonding
Return BTM_CMD_STARTED when we started bonding, as this is expected by
upper layer.
Start service discovery (SDP and GATT) after the remote device is
bonded.
Assume encryption is done when we just bond a device.

TODO: Re-encrypt the link if we connect to a previously bonded device.

Test: cert/run --host
Test: Connect to a LE mouse
Tag: #gd-refactor
Bug: 158861440
Change-Id: I2819702ac068be9515824234e85d81ac2ec7a748