gatt: Add CID for sending notification confirmation
This patch fixes a regression after

commit 7a496c86ba675e1023a5b84de686a823834896db
stack/gatt: Extend calls with CID

Notification confirmation was not taken into account.
For notification function attp_send_cl_msg() is called with p_clcb NULL
which leads to a crash in the following line:

uint16_t payload_size = gatt_tcb_get_payload_size_tx(tcb, p_clcb->cid);

This patch introduces a new function:  attp_send_cl_confirmation_msg()

which is used only for sending confirmation.
Also this patch makes sure that cid is provided to the GATT client so it
can be used in the confirmation call.
Note: handle was removed from the confirmation call as it is redundant.

The only missing part in this patch is handling timers per each
indication and this will be fixed in following patches which will add
eatt support.

Tag: #feature
Bug: 159786353
Sponsor: jpawlowski@
Test: compile & manual

Change-Id: I8bdb38ec673e11744d670223e5416be01666e1a6