Introduce ConfirmationData
ConfirmationData is an object to contain information
for passing to the UI via Display* functions defined
in the class Ui.

This consolidates the required function signature from 2-3 arguments
down to a single argument.  Additionally, we gain a feature
for putting and getting extra data.  This extra data mechanism
will allow for the SecurityModule to pass information to the shim
such as:

    - IO Capability
    - Authentication Requirements
    - MITM bit

It is required that we map these in the shim and pass them back to
BTA in order to drive the BTA state machine properly.

BYPASS_INCLUSIVE_LANGUAGE_REASON="following bluetooth specification terms"

Bug: 162984360
Tag: #gd-refactor
Test: cert/run --host SecurityTest
Change-Id: I30b19fb5a3d5fa3462cf4fdc5f47df5d5bb4c478