L2CAP: Provide a default ERTM option
Note that the existing ERTM options for AVCT_BR and GAP OBEX are almost
the same, and the only difference is TxWindow (10 and 20).  Providing a
default value of 10 won't make a difference.

Use the same ERTM option so that later L2CAP can send config request for
them.

Bug: 159815595
Tag: #refactor
Test: compile & verify basic functions working
Change-Id: I410045f47c643fd4a152ae44722e905b2ce01197