remove invalid test, restucture simple_hal_test
use PyHal abstractions

test_none_event only passed on C++ gd because the reset event
was dropped because the test had not yet asked for the event
stream.

doesn't happen with rust, so the test was failing. by using the
abstractions, the stream gets registered before anthing happens
so the behavior is consistent.

Bug: 171749953
Tag: #gd-refactor
Test: gd/cert/run --rhost SimpleHalTest
Change-Id: I88e7cc45d194030fdff10a247ff8bc09e6473bb5