rusty-gd: initial commit for GDDI
go/gddi for an overview

GDDI (GD dependency injection) aims to solve the issue
of start & stop order, especially in the world where we
bring the stack up partially for testing.

this patch introduces modules & providers, but does not yet save
& remember instances to prevent multiple calls, or provide a way
to seed some object instances (e.g. config), or provide a way to
tear down the started objects

The rest is coming soon.

Bug: 171749953
Tag: #gd-refactor
Test: gd/cert/run --rhost SimpleHalTest
Change-Id: I85f0dac139cfab3174053b92e077c80245da4f85