Security: Wait for name request in link key notification
For classic SSP numeric comparison we must wait at the
UserConfirmationRequest in order to get the name before
we display it to the user.  However, in the Out of Band
data flow we never get a UserConfirmationRequest.

Thus, we must wait in the link key notification if we
have not yet received a name response.

Remove setting to std::nullopt as the operator= in
std::optional causes the underlying pointer to be reset
and thus freed, but still accessed by the pointer we passed.

Bug: 175033589
Test: cert/run SecurityTest:test_successful_dut_initiated_ssp_oob --repeat=20
Tag: #gd-refactor
Change-Id: Ic3f424e22a0c596b21969d4882d0a9b06f7433ab